[PATCH] Merge the "git toolkit" set of shell scripts
This is basically a flat-merge of git-pasky (commit
b44482127ff66768a9bccce478aaa77ea664d485). It consists
of a set of shell scripts which are supposed to abstract
git to be usable like a kind of SCM tool.

These scripts consist mainly of my code, but other code contributors
are in particular:
	Junio C Hamano <junkio@cox.net>
	Ingo Molnar <mingo@elte.hu>
	Christopher Li

Signed-off-by: Petr Baudis <pasky@ucw.cz>