[PATCH] git-pasky spec file
Here's a simple spec file to do rpm builds.  It's against the
latest Makefile (which has the s/BINDIR/bindir/ change).  I've used
DESTDIR, although it's not clear it's meant to stay in the Makefile.
For now, there's no dynamic (git.spec.in, for example) update to the
Version, so it's set against 0.6.3 (expecting it to be forthcoming
shortly).  It installs to /usr/local/bin, and expects the tarball to be
named git-pasky-0.6.3.tar.bz2.  Creates a package named git, which seems
fine since Linus' isn't likely to be packaged directly.  Enjoy.

Signed-off-by: Chris Wright <chrisw@osdl.org>
Signed-off-by: Petr Baudis <pasky@ucw.cz>