[PATCH] git patch, patches are out of order
Running git patch <a>:<b> gives all the patches between versions a and b
in a random order one in which they do not necessarily apply if any
commits modified identical files.

Adding a simple numeric sort on the timestamp to put them in commit order.

Signed-off-by: Jan Harkes <jaharkes@cs.cmu.edu>

Note that they still do not necessarily apply when you take branches into account,
but it's still better than it was before.

Signed-off-by: Petr Baudis <pasky@ucw.cz>