Changed git patch output to match Linus' preferred order for submission, as described in <Pine.LNX.4.58.0504192142040.6467@ppc970.osdl.org>:
You should have "checkin comment" first, then "signed-off-by", then a line
with three dashes, and then "administrative trivia".

Ie I'd much rather see the email look like ...
since otherwise I'll just have to edit it that way. I like seeing the
administrative stuff (diffstat etc), but I don't want to have it in the
commit message, and that's exactly what the "---" marker is for - my tools
will automatically cut it off as if it was a signature (or the beginning
of the patch).