PATCH[4/4]: Fix parent-id mishandling commit text lines starting by 'parent'
parent-id now processes the cat-file output with awk instead of egrep | cut,
making it stop at the empty line terminating the header.

Also added a note to the description that it can return multiple lines of
ids for multiparent commits.

Signed-off-by: Philip Pokorny <ppokorny@mindspring.com>

Dropped the oververbose-commit-id-result-handling bit and rewritten the
commit message of this patch.

Signed-off-by: Petr Baudis <pasky@ucw.cz>