[PATCH] allow cg-log to display commit entries for particular files/directories
This patch allows for one or more paths to be specified with the effect
of restricting the log to commits that affected those files or files
under given directories.

This is not perfect as some merge commits might or might not be listed.
But usually they're not really interesting in the context of this
feature.  It is otherwise pretty simple and effective until a better
tool is available.

For example:

	cg-log arch/arm/kernel/

will show all commits that affected files under arch/arm/kernel/
starting from the head branch.

Signed-off-by: Nicolas Pitre <nico@cam.org>

Simplified since the -r argument was introduced. BTW, I think taking only the
first parent of the merge is the right thing in those cases.

Signed-off-by: Petr Baudis <pasky@ucw.cz>