[PATCH] cogito: replace "git-* $(tree-id)" with "git-* HEAD"
tree-id, when called without parameters, prints the tree ID in the
commit object whose ID is in .git/HEAD.  Both git-diff-cache and
git-read-tree (in general all GIT tools that use get_sha1()) can get
this ID from .git/HEAD themselves when they are called with "HEAD" as
ID parameter.  So this patch replaces occurences of "git-* $(tree-id)"
with "git-* HEAD".  This speeds things up a bit and also makes the
intent of the git-* calls a bit more clear.

Signed-off-by: Rene Scharfe <rene.scharfe@lsrfire.ath.cx>
Signed-off-by: Petr Baudis <pasky@ucw.cz>