[PATCH] have commit-id dereference git tags
With this patch, if commit-id is passed a git tag, it will check to see
if the tag points to a commit. If it does, commit-id will return that
commit's id.

I'm not sure if this is the proper place to handle git tags, but it does
make things like "cg-diff -r v2.6.12-rc2:v2.6.12-rc3" work in the Linus
kernel tree. I've not noticed any problems this introduces, but I didn't
do a thorough check.

Signed-Off-By: Alexander Beyn <malex-git@fatelectrons.org>

Sacked egrep -m 1 (what could -m do?) in favour of simple head -n 1.

Signed-off-by: Petr Baudis <pasky@ucw.cz>