cg-log can now list changed files too
Add an -f switch to cg-log controlling whether to list which files
were changed. The output looks like the following:

        <commit header>

            * file1, file2, ..., fileN:

            <log message>

It doesn't do anything fancy like wrapping long lines and can
be quite costly to do for big logs.

Signed-off-by: Jonas Fonseca <fonseca@diku.dk>

Note that I decided to keep the file list aligned with the log message
since it really should not be considered a fixed well-defined format;
the current format is fine for human reading but sucks for anything
really usable, since ',' can occur in filenames, there's the
leading/trailing crap etc.

Signed-off-by: Petr Baudis <pasky@ucw.cz>