cg-admin-uncommit by default preserves the working tree now
When doing cg-admin-uncommit, it by default now only changes the HEAD
pointer but does not roll back the tree, so the rewound commits show up
as local changes. This is useful for most of the cg-admin-uncommit case
scenarios, when you messed up a commit and just want to redo it. You
can force the old behaviour by passing it a '-t' option.

The implementation is _VERY_ messy now, but I refuse to waste my time
on it now, since this command is not all that important anyway, and it
sorta works.