Support for a postcommit hook
cg-commit will now try to execute

	.git/hooks/commit-post $commit_sha1 $branch_name

after finishing the commit, if the .git/hooks/commit-post hook exists
and is executable.

The hook can also take branch name, since I need to make difference
between different branches. This is solved in an extremely hackish and
ad-hoc way now, just taking .git/branch-name as the branch name if it
exists. We need to figure out whether we want to keep it this way or
figure out something better.