README rewrite
This patch is a complete rewrite of the Cogito section of the README file.
Commands are explained in detail, with a quickstart section at the top and a
full command reference below.

Signed-off-by: Zack Brown <zbrown@tumblerings.org>

I've actually used only the "introduction" part for now, the reference part
should live in the separate files of Documenation/ directory. I also
substantially tweaked it.

Signed-off-by: Petr Baudis <pasky@ucw.cz>