[PATCH] Make cg-status report added/removed files correctly
cg-status presently only cares about the cache (by parsing the output of
git-update-cache and git-ls-files), when it should look at the tree in HEAD.

This patch fixes this, by using git-diff-cache HEAD instead (using
git-diff-files only to detect the difference between missing files and
cg-rm'd ones).

It also uses the exclusion mechanism in git-ls-files, instead of parsing
git-ls-files output, and adds '*~' and '#*' to the list of patterns
excluded by default.

Signed-off-by: Dan Holmsand <holmsand@gmail.com>
Signed-off-by: Petr Baudis <pasky@ucw.cz>