Fixed cg-log, cg-mkpatch and cg-log argument processing
My 'improvement' of the parsing loop by introducing $arg obviously broke
it since we would lose one argument after the dash-arguments (the file
to focus on).

The cg-diff fix was more complex since we lost the filtering mechanism
while abandoning cg-Xdiffdo in favour of git-diff -p. Luckily it can
take the pathspec as the trailing arguments.

Reported by Marcel Holtzmann.