[PATCH] fix cg-commit new file handling
There is a problem with committing new files with the current cogito
version.

I got an error like this:
	fatal: Unable to add arch/ppc/kernel/dcr.S to database

I think this is caused by the incorrect file list parsing in
cg-commit precommit_update(): it expects new files prefixed with "A",
while they are prefixed with "N".

The following trivial patch fixes this.

Signed-off-by: Eugene Surovegin <ebs@ebshome.net>
Signed-off-by: Petr Baudis <pasky@ucw.cz>