[PATCH 1/6] Add infrastructure for option parsing to cg-Xlib
Adds a new function, optparse, that handles combined options
just like optget (cg-log -cf, for example).

When called without any argument, it returns 0 if there is any
option to parse.

When called as "optparse -f" or "optparse --foo" it returns 0
if the current option in $ARGS (at $ARGPOS) matches.

When called as "optparse -f= or "optparse --foo=" it requires
an argument.

For long options, it takes an optional second argument, specifying
the number of characters in the option name that has to be present
(defaults to 1). So "optparse --foo 2" will match "--fo", but not
"--f".

Just as with optget, options can be given after arguments as
well, as in "cg-log Documentation -rorigin". Option parsing
stops after "--".

optparse leaves unparsed arguments in the $ARGS array.

Signed-off-by: Dan Holmsand <holmsand@gmail.com>

Tidyups.

Make cg-log and cg-diff use the new optparse stuff.

Signed-off-by: Petr Baudis <pasky@ucw.cz>