[PATCH 3/6] Improve the showdate date displaying routine
This simplifies the showdate routine, and perhaps even increases
portability a little.

Note that this changes the calling convention: the previous
version used seconds from $1, but timezone from the global
variable $date. The users are modified to the new way.

Also fixes bash's belief that number literals starting with
zero are octal.

Signed-off-by: Dan Holmsand <holmsand@gmail.com>

Fixed missing changes in cg-log and dropped the change to Linus format
(the format sucks).

Signed-off-by: Petr Baudis <pasky@ucw.cz>