Commit abortion prompt
cg-commit now offers a helpful confirmation prompt in case you abort the
commit message edit. It will offer you to commit anyway (in case the
default commit message was good enough), and in case you still insist
on aborting, it will inform you that the merge is not aborted if there
was any in progress - that was confusing people.