[Cogito] Less verbose cg-clone/cg-update (cg-pull, in fact)
Here is a patch that implements a cogito-based "verbose" output option (-v)
for cg-pull.  The default is to show only the "status" file change letter
and the file name in the tree delta, with -v showing full git-diff-tree
output.  Also, with the option in there, it wouldn't be hard to quiet up
other portions of cg-pull.

Signed-off-by: Frank Sorenson <frank@tuxrocks.com>

Changed from -q to -v and inverted.

Signed-off-by: Petr Baudis <pasky@ucw.cz>