Deprecate the rsync repository access method
Rsync is bad. Worse if you buy the packed files etc, and the 'git-only'
access methods are growing mature, so remove any example references
to the rsync protocol and make cg-pull warn about the fact that it's
deprecated. I will likely remove it after the next one or few releases.