Make cg-tag create the tag objects only if you pass it -s
I changed my mind. In many cases the cryptographic signature is useless,
and so is then the tag object. So cg-tag without -s will behave as
before I introduced the tag object support and just make it reference
the commit object. With -s, cg-tag will create the tag object and always
try to sign it.