Fix cg-pull to handle packed tags properly
If the objects referenced by refs/tags/ are packed, it wouldn't detect
them properly and instead try to refetch them, but they are likely to
be packed on the other side as well and that makes them impossible to
be fetched explicitly (which isn't a problem as long as they are the
same branch).

Also, the fetch failure message was confusing.

Reported by Russel King.