Add support for .gitignore into cg-status
cg-status will now ignore files matched by .gitignore along the path
to the subdirectory it gets ran from. Well, it now can be called only
from the root directory so that's moot, but it works that way and I even
tested it.

I chose .gitignore since I think this should become a pan-git
convention. That can change yet based on some ML discussion, but I hope
it to stay that way.