More robust patch processing in cg-patch
cg-patch's patch parser is now much more robust and of simpler code. It
no longer requires the ---, +++ lines to be present for a patch to
be properly parsed (not the case for pure mode changes; before, the
modechanges etc. would be applied at the time +++ line is hit), and now
takes the filenames from the diff --git line.

This should finally solve Wolfgang Denk's troubles.