[PATCH] Debianize Cogito
attached is a patch containing a debianziation update for cogito, based
upon the Debian/unstable package.  I mailed the official Debian
maintainer in the hope to get an updated .deb but got no reply from him
until today.  So i used his work to make a new release of it.

Signed-off-by: Holger Eitzenberger <holger@my-eitzenberger.de>

Note that the build itself does not work yet since I turned down the
Makefile changes so the build script needs to be fixed accordingly. This
builds upon git's debian/ directory.

Signed-off-by: Petr Baudis <pasky@ucw.cz>