Unbundle Linus'/Junio's GIT sources from Cogito
GIT is no longer part of Cogito distribution. It appears as it is
hopefully mature enough by now that it will keep backwards compatibility
and will follow a regular release schedule. This should make it a
lot easier for distributions to provide us as well as for system
administrators to keep us up to date (especially on kernel.org ;-).

This commit removes all GIT-related files from the tree and modifies
other relevant files so that they are Cogito-specific only, and obviously
updates the Makefiles.