[PATCH] -x (no-exclude) option for cg-status
This patch adds the "-x" option that instructs cg-status not to ignore
any files.  Also, the code has been changed to add standard includes in
a loop.

I'm trying to reimplement my cg-clean script so that it reuses more code
from other utilities.  This patch makes it possible to call cg-status
from cg-clean.

Signed-off-by: Pavel Roskin <proski@gnu.org>
Signed-off-by: Petr Baudis <pasky@ucw.cz>