Support cg-clone -l, which symlinks the repositories
Cloning locally is awfully slow normally, since all the objects are
hardlinked. This is a safe bet and should survive mostly everything,
but bogs the cloning down and makes it slow, while frequently what is
enough is symlinking the object repository and that's it. Now you can
do that by passing cg-clone a '-l' argument. You can also do it anytime
later, cg-pull and cg-push will detect that automatically (so it should
also deal properly with pulling in the other direction, pushing and such).