[PATCH] Identical trees should not trigger error in cg-diff
cg-diff reports error when comparing identical trees.  It's not as
useless as it may seem.  For example, I make local changes, e-mail them
upstream and then get them back via cg-update with minor changes.  I
revert the minor differences between master and origin.  Here's what I
get when I want to make sure the trees are identical (their history is
not):

	$ cg-diff -r origin:master
	cg-diff: trying to diff b00f462cee5b02455abdbb476fa7c9e94a9fcb4c
	against itself

What's worse, cg-diff exists with exit status 1, indicating an error.  I
don't see any reason for this to be an error.  No difference between
branches is OK.  This may be especially important when cg-diff is used
non-interactively in a script.

Signed-off-by: Pavel Roskin <proski@gnu.org>
Signed-off-by: Petr Baudis <pasky@suse.cz>