[PATCH] cg-clean -d removes "arch" and "include" in Linux
"cg-clean -d" would delete directories that don't have versioned files
in them, even if their subdirectories do.  "arch" and "include" in Linux
sources are such directories.

To be safe, every directory in the directory list should be complemented
by all its parent directories.

The patch below is a quick hack that shows that bash is not the best
language to implement this.  Still, it's better to have such fix than
having cg-clean remove source files.

Signed-off-by: Pavel Roskin <proski@gnu.org>
Signed-off-by: Petr Baudis <pasky@suse.cz>