[Terminology] pull -> fetch; especially cg-pull -> cg-fetch
Use 'fetch' everywhere we would use 'pull' instead before. Most notably
rename cg-pull to cg-fetch.

I believe such a change so late in the process is a bit unfortunate,
but I was persuaded that consistency with GIT is really important.

Some time after the pull alias will be removed, cg-update will be
aliased to cg-pull, since 'pull' now really means 'update' in the GIT
world. This only makes the situation worse and more confusing for the
users, of course.