Fix the default-to-origin code
We would default to origin only when the origin ref already actually
exists, but that may not be the case when we are doing initial origin
pull. Normally we would always already have the ref because we cloned,
but perhaps we init'd the repository and added the origin branch later
just because we pull from another repository very frequently. So we now
default to origin just if .git/branches/origin exists.