Major cg-restore revamp - restoring from history and of removed files
The code in cg-restore triggerred when specific files are passed to it
was totally rewritten. cg-restore now accepts -r argument which enables
you to 'restore' your working copy file version to a historical one,
either in a given commit or even just to replace the file with given blob.

As a consequence, cg-restore -f <file-removed-with-cg-rm> will readd
the file.