cg-merge -j for joining disjoint branches
Sometimes you want to join two projects together, and that's where
cg-merge -j hurries to help. It will work _only_ in that case and without
it, cg-merge will refuse to do anything, so it works like some kind of
'yes, I really want to do it in this case', since sometimes it may be
just a plain user error.

Inspired by Alan Chandler.