Proper tags fetching
From now on, tags are fetched 'properly'. That means we ask git-ls-remote
for their list, which requires server info for HTTP, but it enables us
to fetch the tag only when we have the object it ultimately references,
making this autotagsfetching much more correct (although it won't be
entirely correct until we get private tags support).

Inspired by snippet posted by Junio, but heavily modified.