Fix the infamous fastforwarding stale files bug
When fastforwarding (or seeking, for that matter), sometimes some files
wouldn't be removed although they should've been. I didn't actually got
to the root of the bug, but it is something extremely weird seemingly,
but not entirely, connected with the maximal argument count and some
xargs and especially bash weirdness.

This changes the code to use much saner pipe communication instead. We
don't support newlines in files anymore, but that's no big deal since
we don't in the rest of Cogito either.