[PATCH] Add short help to Cogito commands
Captions for all 74 Cogito options are added to the script headers on
the same line as the option being documented. The preferred syntax is
now:

	# <option>:: <caption>
	#       <description>

The old help message (the complete user manual) is moved to --long-help
and -h and --help both will now show the short help message. Output of
cg-help <cmd> is untouched. In the progress, rewrite cg-Xlib's
print_help to use one built-in type -P instead of multiple which's.

Documentation scripts (to generate cg-manpages and the long help
messages) has been updated to ignore the option captions.

Missing description of cg-admin-ls -t option was added.

Signed-off-by: Jonas Fonseca <fonseca@diku.dk>
Signed-off-by: Petr Baudis <pasky@suse.cz>