Initial cg-push to a remote branch wouldn't work properly
While a while ago we modified cg-fetch, cg-merge, and cg-update to default
to origin just when .git/branches/origin exists and .git/refs/heads/origin
does not have to exist, we forget to do the same for cg-push - so in case
there was no head for origin yet, it wouldn't default to origin. This
breaks when you do the first push ever to the remote repository (or any
further pushes until you fetch the next time, for that matter, since we
don't update the head when pushing - we should, though).