cg-switch now requires -r when creating new branch
As someone already suggested, cg-switch -r being required for new branches
is useful in case you make a typo when switching to another branch -
a new branch would be created, which is not what you want, and this one
is really easy to do. This requirement makes it more robust, and you
don't create new branches every day, so it shouldn't be a big hassle.

While at it, improve the wording of the relevant documentation a bit.