Whitespaces safety fixes
Fix hopefully all the possible whitespace and metacharacters related
problems by carefully quoting everything.  Even most of the things clearly
unneeded, because it doesn't harm and I want to cultivate a reflex to panic
and closely investigate if anything unquoted passes by, so that we _stay_
safe in the future as well.

Such a large patch is bound to introduce some (probably mostly trivial) bugs,
so please test this carefully.