[PATCH] cg-update quotes too much
It seems that cogito commit e5ca051c1d3900f7fbc1592d018ab3aab6a9573a
broke cg-update, as expected ("Such a large patch is bound to introduce
some (probably mostly trivial) bugs").

Quoting $force and $squash gives cg-fetch and cg-merge an empty argument
if they are not set. If they are not quoted they will be omitted if they
are empty, which is the right thing.

Signed-off-by: Tero Roponen <teanropo@cc.jyu.fi>
Signed-off-by: Petr Baudis <pasky@suse.cz>