[PATCH] Fix unconditional early exit in cg-fetch v2
When invoking cg-fetch, after fetching tags it exits here, for an overlooked
error handling. Replace the exit with an early return.

This means, for instance, we exit without reporting the tag updates and without
removing the "fetch in progress" marker - leading to unconditional "Recovering
from interrupted fetch" at the very beginning with rsync transport. Indeed, this
is fixed by this patch.

Signed-off-by: Paolo 'Blaisorblade' Giarrusso <blaisorblade@yahoo.it>
Signed-off-by: Petr Baudis <pasky@suse.cz>