Rewrite the cutesy fetch progressbar
The main change is that it will show ongoing large file (packs) download
progress now, and there are some other minor cosmetic changes.

This sucks for rsync and will problably suck on - rsync uses buffered
i/o when not talking to a terminal, so we are not getting continuous
stream of up-to-date information.

Inspired by Pavel Machek's complaints. ;)