Resumable cg-clone - if interrupted, resume by cg-fetch
This patch moves the initial checkout functionality to cg-fetch and
makes the initial-clone information persistent, while also removing
the cg-clean cleanup hook. This means that when the initial fetch fails
for some reason, the repository is not deleted and you can restart the
initial clone by just running cg-fetch.

Pretty neat, if you ask me.