[PATCH] Make cogito work with accented letters
git-diff-index in default mode has an annoying behaviour wrt filenames
containing non-ascii chars.  As suggested by Pasky, we can use -z
mode, which gives us a much better way of handling all other special
chars.  With associated testcases ensuring it works with simple and
double quotes, backslashes, and spaces as well.

This is an improved version of the previous patch, which fixes other
commands like cg-switch which use tree_timewarp.  A couple of other
commands using git-diff-index still need to be fixed in a similar
manner.

Signed-off-by: Yann Dirson <ydirson@altern.org>
Signed-off-by: Petr Baudis <pasky@ucw.cz>