[PATCH 1/2] Fix cg-patch hanging on terminals with TOSTOP flag
Terminals with TOSTOP flag don't allow background processes to write to
the terminal.  This flag is used e.g. by the subshell in GNU Midnight
Commander.  It can also be set by "stty tostop".

Redirect stdout and strerr of the "patch" command to a temporary file
and show it once "patch" terminates.

Signed-off-by: Pavel Roskin <proski@gnu.org>
Signed-off-by: Petr Baudis <pasky@ucw.cz>