[PATCH] cg-admin-rewritehist: Support multiple parents of the start revision (-r).
This uses cg-object-id -p, which resolves symbolic commit names for us
into sha1s.

The construct ${unchanged{@}/#/^} replaces the empty string at the beginning (#)
of each element in the array 'unchanged' by a ^ and returns a list of the so
constructed words (i.e. ^sha1 ^sha2 ...). It is important not to quote this.

Signed-off-by: Johannes Sixt <johannes.sixt@telecom.at>
Signed-off-by: Petr Baudis <pasky@suse.cz>