Be way more paranoid when failing to apply patches w/ cg-commit -p
Resolving the situation to a sane state is not that simple to do
automagically, so that'll be done in a separate commit. For now just do
not delete anything, tell the user about where to find all the relevant
files and let her somehow resolve the messup.

Breakage pointed out by Dennis Stosberg <dennis@stosberg.net>.