Do three-way merge when timewarping tree with local changes
Currently, the local changes would be stashed to a diff, erased, then the
tree timewarped and then the local changes applied back. This was prone
to failure if something bad happenned when the diff was stashed away,
when the changes did something cg-patch did not deal well with or when
conflicts happenned - in that case you would have to live with .rej files,
furthermore overwritten by _further_ timewarps. Bad, bad, bad.

This is one of the most long-standing problems with Cogito, hopefully
finally going away now. :-) Please test. (Well, except that noone will
read this and if anyone will, they will not bother to test anything
anyway. Oh, what the hell.)