Do not pass cg-commit the -E in cg-init
I think this was a mistake to do in the first place. Being consistent
is more important than having 'nicer' handling of inits, besides this
does give you a chance to abort the init (well, not really, but that'll
be fixed next). Besides, it's now broken if input is not a tty. ;-)