Implement cg-commit --push for autopushing commits to remote branches
in a transactional way - that is, if the push fails, the commit is
not recorded (but you can reuse the commit message at the next try
easily). This implements the CVS-ish update-before-commit model which is
really inferior and you shouldn't actually _use_ this, but there might
be cases where you could want to do it anyway. Besides, the Bazaar folks
do this as the 'heavyweight checkout'. ;-)