Revert 733694492f13 (default LESS=F)
This breaks in X terminals with alternate screen turned on, the output
is immediatelly cleared as the terminal switches back to the main screen.
Use either LESS=FX in those terminals to disable the alternate screen,
or LESS=F if the alternate screen is irrelevant for you (e.g. linux
console).