cg-merge: Fix following volatile branches
It's been all wrong, but it would be visible only with -v, or when
preserving local branches, and since it happens rarely anyway, noone
probably noticed. Only the new timewarping mechanism uncovered the bug.

We must use $head, not $base, as the source commit. Think about it. $base
is not interesting at all since we are moving from the current _head_,
so that's what the difference between should be also shown. $base is
implementation detail of whoever rebased the volatile branch.