Make cg-patch a git-apply frontend
This drops a whole lot of pretty awful code in favour of just using
git-apply which seems to be finally actually usable for fuzzy patch
application and reverting. I hope this creates no weird problems for
people.

This also drops the -u option to cg-patch, but it was such a nuance that
I don't think anyone's ever really been using it.